Chapter 1
The movement of financial assets and products generally displays some type of expected return, even over a short period of time. This expected return trends at a predictable rate that may be positive indicating growth, negative indicating a decline, or zero. Additionally, there are random movements that are individually unpredictable; however, the general distribution of these fluctuations are predictable based on historical movements. The common approach to model randomness is to assume a single or multi-component Gaussian process.